Minor fixes for the type defination
